refactor(html): tokenization state: character reference
